backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# Redis
redis[hiredis]==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# AI
groq==0.8.0

# Configuration
pydantic-settings==2.0.3

# Validation
email-validator==2.1.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
requests==2.31.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Environment
python-dotenv==1.0.0
